+++ cmd/provisioner.go
@@ -2,15 +2,40 @@ package cmd
 
 import (
 	"github.com/spf13/cobra"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
 )
+
+var provisionerFlags = struct {
+	k8sOverrides clientcmd.ConfigOverrides
+
+	restConfig *rest.Config
+	namespace  string
+}{}
 
 var provisionerCmd = &cobra.Command{
 	Use:   "provisioner",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
+	Short: "Provision workers that runs builds",
+	Long: `Provisions workers, which are Kubernetes pods running wharf-cmd
+that clones the repository and run the .wharf-ci.yml file inside the repo.
+
+The "wharf-cmd provisioner" act as a fire-and-forget, where the entire build
+orchestration is handled inside the Kubernetes cluster, in comparison to the
+"wharf-cmd run" command that uses your local machine to orchestrate the build.
 `,
+	PersistentPreRunE: func(cmd *cobra.Command, args []string) error {
+		restConfig, ns, err := loadKubeconfig(provisionerFlags.k8sOverrides)
+		if err != nil {
+			return err
+		}
+		provisionerFlags.restConfig = restConfig
+		provisionerFlags.namespace = ns
+		return nil
+	},
 }
 
 func init() {
 	rootCmd.AddCommand(provisionerCmd)
+
+	addKubernetesFlags(provisionerCmd.PersistentFlags(), &provisionerFlags.k8sOverrides)
 }