+++ cmd/provisioner_create.go
@@ -9,11 +9,12 @@ import (
 
 var provisionerCreateCmd = &cobra.Command{
 	Use:   "create",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
+	Short: "Starts a build via a new worker inside a Kubernetes pod",
+	Long: `Creates a new Kubernetes pod that clones a Git repo and
+a container running "wharf-cmd run" to perform the build.
 `,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		p, err := provisioner.NewK8sProvisioner("default", Kubeconfig)
+		p, err := provisioner.NewK8sProvisioner(provisionerFlags.namespace, provisionerFlags.restConfig)
 		if err != nil {
 			return err
 		}