+++ cmd/run.go
@@ -8,24 +8,44 @@ import (
 	"github.com/iver-wharf/wharf-cmd/pkg/wharfyml"
 	"github.com/iver-wharf/wharf-cmd/pkg/worker"
 	"github.com/spf13/cobra"
+	"k8s.io/client-go/tools/clientcmd"
 )
 
-var flagRunPath string
-var flagStage string
+var runFlags = struct {
+	path         string
+	stage        string
+	env          string
+	k8sOverrides clientcmd.ConfigOverrides
+}{}
 
 var runCmd = &cobra.Command{
 	Use:   "run",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
+	Short: "Runs a new build from a .wharf-ci.yml file",
+	Long: `Runs a new build in a Kubernetes cluster using pods
+based on a .wharf-ci.yml file.
+
+If no stage is specified via --stage then wharf-cmd will run all stages
+in sequence, based on their order of declaration in the .wharf-ci.yml file.
+
+All steps in each stage will be run in parallel for each stage.
+
+Read more about the .wharf-ci.yml file here:
+https://iver-wharf.github.io/#/usage-wharfyml/
 `,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		stepRun, err := worker.NewK8sStepRunner("build", Kubeconfig)
+		kubeconfig, ns, err := loadKubeconfig(runFlags.k8sOverrides)
+		if err != nil {
+			return err
+		}
+		stepRun, err := worker.NewK8sStepRunner(ns, kubeconfig)
 		if err != nil {
 			return err
 		}
 		stageRun := worker.NewStageRunner(stepRun)
 		b := worker.New(stageRun)
-		def, errs := wharfyml.ParseFile(flagRunPath)
+		def, errs := wharfyml.ParseFile(runFlags.path, wharfyml.Args{
+			Env: runFlags.env,
+		})
 		if len(errs) > 0 {
 			log.Warn().WithInt("errors", len(errs)).Message("Cannot run build due to parsing errors.")
 			for _, err := range errs {
@@ -40,7 +60,7 @@ var runCmd = &cobra.Command{
 			return errors.New("failed to parse .wharf-ci.yml")
 		}
 		res, err := b.Build(context.Background(), def, worker.BuildOptions{
-			StageFilter: flagStage,
+			StageFilter: runFlags.stage,
 		})
 		if err != nil {
 			return err
@@ -59,6 +79,8 @@ var runCmd = &cobra.Command{
 func init() {
 	rootCmd.AddCommand(runCmd)
 
-	runCmd.Flags().StringVarP(&flagRunPath, "path", "p", ".wharf-ci.yml", "Path to .wharf-ci file")
-	runCmd.Flags().StringVarP(&flagStage, "stage", "s", "", "Stage to run (will run all stages if unset)")
+	runCmd.Flags().StringVarP(&runFlags.path, "path", "p", ".wharf-ci.yml", "Path to .wharf-ci file")
+	runCmd.Flags().StringVarP(&runFlags.stage, "stage", "s", "", "Stage to run (will run all stages if unset)")
+	runCmd.Flags().StringVarP(&runFlags.env, "environment", "e", "", "Environment selection")
+	addKubernetesFlags(runCmd.Flags(), &runFlags.k8sOverrides)
 }